requirements.txt
django
djangorestframework
markdown
django-filter
cryptography
django-rest-knox
python-decouple